std: Move array_each out of ptr
